+++ src/main/kotlin/betting/BetResult.kt
@@ -4,9 +4,9 @@ data class BetResult (
     val bet: Bet,
     val winning: Winning = Winning(),
 ) {
-    fun win(amount: Long?): BetResult = BetResult(bet, Winning(amount = amount ?: 0L))
+    fun win(amount: Double?): BetResult = BetResult(bet, Winning(amount = amount ?: 0.0))
 
     fun lose(): BetResult = BetResult(bet, Winning(amount = bet.negative()))
 
-    fun lose(amount: Long?): BetResult = BetResult(bet, Winning(amount = this.winning.amount + (amount ?: 0L)))
+    fun lose(amount: Double?): BetResult = BetResult(bet, Winning(amount = this.winning.amount + (amount ?: 0.0)))
 }
